:recycle: refactor(PatientPicker): Extract usePatientsSuggestions.
