+++ imports/ui/patients/PatientPicker.js
@@ -1,45 +1,17 @@
 import React from 'react';
 
-import {patients, usePatientsAdvancedFind} from '../../api/patients.js';
-import {normalizeSearch} from '../../api/string.js';
+import {patients} from '../../api/patients.js';
 
 import SetPicker from '../input/SetPicker.js';
 
-const useSuggestions = (searchString) => {
-	const $search = normalizeSearch(searchString);
-	const limit = 5;
-
-	const query = {$text: {$search}};
-
-	const sort = {
-		score: {$meta: 'textScore'}
-	};
-	const fields = {
-		...sort,
-		_id: 1,
-		firstname: 1,
-		lastname: 1
-	};
-
-	const options = {
-		fields,
-		sort,
-		skip: 0,
-		limit
-	};
-
-	return usePatientsAdvancedFind(query, options, [
-		$search
-		// refreshKey,
-	]);
-};
+import usePatientsSuggestions from './usePatientsSuggestions';
 
 const PatientPicker = ({...rest}) => {
 	return (
 		<SetPicker
 			itemToKey={patients.toKey}
 			itemToString={patients.toString}
-			useSuggestions={useSuggestions}
+			useSuggestions={usePatientsSuggestions}
 			maxCount={1}
 			{...rest}
 		/>
